Fix compile error

+++ src/github.com/indyjo/bitwrk-client/trade.go
@@ -421,7 +421,7 @@ func (t *Trade) GetState() *ActivityState {
 	}
 
 	if t.workFile != nil {
-		result.BytesTotal = uint64(t.workFile.Size())
+		result.BytesTotal = t.workFile.Size()
 	}
 
 	return result
